requirements.txt
python-dotenv==1.0.1
python-telegram-bot[job-queue]==21.7
asyncpg==0.30.0
httpx[http2]==0.28.1
pydantic-settings==2.6.1
apscheduler==3.10.4
pytest==8.3.4
pytest-asyncio==0.24.0
